+++ back/src/main/java/com/FindMyPc/back/serviceImpl/WishlistServiceImpl.java
@@ -1,13 +1,16 @@
 package com.FindMyPc.back.serviceImpl;
 
+import com.FindMyPc.back.ResponseDto.WishlistResponseDto;
 import com.FindMyPc.back.entity.Wishlist;
 import com.FindMyPc.back.repository.WishlistRepository;
 import com.FindMyPc.back.service.WishlistService;
+import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
 import java.util.Optional;
+import java.util.stream.Collectors;
 
 @Service
 public class WishlistServiceImpl implements WishlistService {
@@ -15,29 +18,36 @@ public class WishlistServiceImpl implements WishlistService {
     @Autowired
     private WishlistRepository wishlistRepository;
 
+    @Autowired
+    private ModelMapper modelMapper;
+
     @Override
-    public List<Wishlist> getAllWishlists() {
-        return wishlistRepository.findAll();
+    public List<WishlistResponseDto> getAllWishlists() {
+        List<Wishlist> wishlists = wishlistRepository.findAll();
+        return wishlists.stream()
+                .map(wishlist -> modelMapper.map(wishlist, WishlistResponseDto.class))
+                .collect(Collectors.toList());
     }
 
     @Override
-    public Optional<Wishlist> getWishlistById(int id) {
-        return wishlistRepository.findById(id);
+    public Optional<WishlistResponseDto> getWishlistById(int id) {
+        Optional<Wishlist> wishlistOptional = wishlistRepository.findById(id);
+        return wishlistOptional.map(wishlist -> modelMapper.map(wishlist, WishlistResponseDto.class));
     }
 
     @Override
-    public Wishlist saveWishlist(Wishlist wishlist) {
-        return wishlistRepository.save(wishlist);
-    }
-    
-    @Override
-    public Wishlist updateWishlist(Wishlist wishlist) {
-        return wishlistRepository.save(wishlist);
+    public WishlistResponseDto saveWishlist(WishlistResponseDto wishlistResponseDto) {
+        Wishlist wishlist = modelMapper.map(wishlistResponseDto, Wishlist.class);
+        Wishlist savedWishlist = wishlistRepository.save(wishlist);
+        return modelMapper.map(savedWishlist, WishlistResponseDto.class);
     }
 
 
     @Override
     public void deleteWishlist(int id) {
-        wishlistRepository.deleteById(id);
+        if (wishlistRepository.existsById(id)) {
+            wishlistRepository.deleteById(id);
+        }
+        // Handle the case where the wishlist does not exist
     }
 }
